Add article alias in building article SEF route when missing
Like com_content router does, when invoking JRoute on an article view, if just the id is passed and not the full slug (id:alias), the component router should add the alias. This is important since when SEF is the default "id-alias", it should not be mandatory to write a link with the alias since it can be changed in the future and that will produce a broken link in Seblod with 404 Error.
This PR retrieve article alias from #__content and adds it to the article id if the alias is missing. 
@sebastienheraud: can you check if this addition is correctly placed in "$config['doSEF'] == 'full'" condition? thanks